modified while clause in divide
Initially I had just pow(2,level) as the upper limit for the while loop but realised this fails when the number of elements in the list exceeds a certain number.
So I added a constraint min( pow, length_of_input_array ). This will limit the search to favour the result.
Tested with some basic examples. Need to test thoroughly.