minor bug fix (see expln)
The code was not printing a point as a convex hull point when it belonged to the vertical boundary.
Hence in the mergeConvexHulls, changed weak inequality to strict inequality.
Also had a rubbish piece of code (with console stt like THIS WILL NOT GET EXED...) removed.
Now it seems to be working fine. Will integrate with the divide code and see.