sanity check after modifying BT (check desc)
so I have added the inorder and preorder inputs in the previous commit.
I tried to redraw the nodes in black colour when the execution has shifted to previous level in conquer.
It is giving error. hence that part is commented out.
It displays the node correctly in black colour but the arros/lines get messed up.
I suspect it is the getPointInOuterSVG() function that is causing the error. Not sure though.
Will check and fix it tom..oops no today as it is 0201 hrs BST!